Monitoring opened files for changes outside BASIC256 editor.
Alerts user when a file has been modified, renamed or removed from disk (outside of BASIC256 editor).
If file is renamed or removed the user get an warning message and the opened file is marked as not saved to avoid loosing work when user close the application.
If file is modified from outside of BASIC256 editor, then the user is asked if wants to reload the modified file. If user choose to load the new file, then the UNDO history is not cleared to allow user to revert loading file operation.

git-svn-id: https://svn.code.sf.net/p/kidbasic/code/trunk@884 deda80ff-9d1e-0410-aade-8f0ad8a54eb3